requirements.txt
opencv-python
matplotlib
keras-nightly
numpy
tensorflow
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
pydantic>=2.4.0